+++ scripts/postprocessing_upscale.py
@@ -10,6 +10,7 @@ from modules.processing import apply_color_correction, setup_color_correction
 from modules.ui import switch_values_symbol
 from modules.ui_common import create_refresh_button
 from modules.ui_components import FormRow, InputAccordion, ToolButton
+from modules.upscaler_ui import upscaler_choices
 
 upscale_cache = {}
 
@@ -50,9 +51,9 @@ class ScriptPostprocessingUpscale(scripts_postprocessing.ScriptPostprocessing):
                                 upscaling_crop = gr.Checkbox(label="Crop to fit", value=True, elem_id="extras_upscaling_crop")
 
             with FormRow():
-                extras_upscaler_1 = gr.Dropdown(label="Upscaler 1", elem_id="extras_upscaler_1", choices=[x.name for x in shared.sd_upscalers], value=shared.sd_upscalers[0].name)
-                extras_upscaler_2 = gr.Dropdown(label="Upscaler 2", elem_id="extras_upscaler_2", choices=[x.name for x in shared.sd_upscalers], value=shared.sd_upscalers[0].name)
-                create_refresh_button([extras_upscaler_1, extras_upscaler_2], load_upscalers, lambda: {"choices": [x.name for x in shared.sd_upscalers]}, "refresh_upscaler")
+                extras_upscaler_1 = gr.Dropdown(label="Upscaler 1", elem_id="extras_upscaler_1", choices=upscaler_choices(shared.sd_upscalers), value=shared.sd_upscalers[0].name)
+                extras_upscaler_2 = gr.Dropdown(label="Upscaler 2", elem_id="extras_upscaler_2", choices=upscaler_choices(shared.sd_upscalers), value=shared.sd_upscalers[0].name)
+                create_refresh_button([extras_upscaler_1, extras_upscaler_2], load_upscalers, lambda: {"choices": upscaler_choices(shared.sd_upscalers)}, "refresh_upscaler")
 
             with FormRow():
                 extras_color_correction = gr.Checkbox(label="Color Correction", elem_id="extras_color_correction", value=False)